introduced ssemanagers.go
